+++ spec/workbook/pivot_table/pivot_table_location_spec.rb
@@ -1,0 +1,28 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe OoxmlParser::PivotTableDefinition do
+  let(:xlsx) { OoxmlParser::Parser.parse('spec/workbook/pivot_table/pivot_table.xlsx') }
+  let(:location) { xlsx.pivot_table_definitions[0].location }
+
+  it 'location is correct class' do
+    expect(location).to be_a(OoxmlParser::Location)
+  end
+
+  it 'location ref is correct' do
+    expect(location.ref).to eq('A3:C20')
+  end
+
+  it 'location first_header_row is correct' do
+    expect(location.first_header_row).to eq(true)
+  end
+
+  it 'location first_data_row is correct' do
+    expect(location.first_data_row).to eq(true)
+  end
+
+  it 'location first_data_column is correct' do
+    expect(location.first_data_column).to eq(false)
+  end
+end