+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Add support of parsing `CacheFields#cache_field`
 * Add support of parsing `CacheField#shared_item`
 * Add support of parsing `PivotTableDefinition` properties
+* Add support of parsing `PivotTableDefinition#location`
 
 ### Changes